# server/api/openapi-with-docs.json
{
  "x-generator": "NSwag v14.6.3.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Command/{deviceId}": {
      "post": {
        "tags": [
          "Command"
        ],
        "operationId": "Command_SendCommand",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "command",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameCommandDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/WebApi/whackabird": {
      "get": {
        "tags": [
          "WebApi"
        ],
        "operationId": "WebApi_GetWhackabirdScores",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeListenResponseOfListOfWhackabirdscore"
                }
              }
            }
          }
        }
      }
    },
    "/api/WebApi/redbirdgreenbird": {
      "get": {
        "tags": [
          "WebApi"
        ],
        "operationId": "WebApi_GetRedbirdScores",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeListenResponseOfListOfRedbirdgreenbirdscore"
                }
              }
            }
          }
        }
      }
    },
    "/api/WebApi/birdiesays": {
      "get": {
        "tags": [
          "WebApi"
        ],
        "operationId": "WebApi_GetBirdieSaysScores",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeListenResponseOfListOfBirdiesaysscore"
                }
              }
            }
          }
        }
      }
    },
    "/api/WebApi/sse": {
      "get": {
        "tags": [
          "WebApi"
        ],
        "operationId": "WebApi_Connect",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GameCommandDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "game": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "playerName": {
            "type": "string"
          }
        }
      },
      "RealtimeListenResponseOfListOfWhackabirdscore": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RealtimeListenResponse"
          },
          {
            "type": "object",
            "description": "Returned by subscribe endpoints with initial data. The client receives the current state\nimmediately and knows which SSE group to listen on for subsequent updates.",
            "additionalProperties": false,
            "properties": {
              "data": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/Whackabirdscore"
                }
              }
            }
          }
        ]
      },
      "Whackabirdscore": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "playerName": {
            "type": "string"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "RealtimeListenResponse": {
        "type": "object",
        "description": "Returned by subscribe endpoints so the client knows which SSE group to listen on.",
        "additionalProperties": false,
        "properties": {
          "group": {
            "type": "string"
          }
        }
      },
      "RealtimeListenResponseOfListOfRedbirdgreenbirdscore": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RealtimeListenResponse"
          },
          {
            "type": "object",
            "description": "Returned by subscribe endpoints with initial data. The client receives the current state\nimmediately and knows which SSE group to listen on for subsequent updates.",
            "additionalProperties": false,
            "properties": {
              "data": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/Redbirdgreenbirdscore"
                }
              }
            }
          }
        ]
      },
      "Redbirdgreenbirdscore": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "playerName": {
            "type": "string"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "RealtimeListenResponseOfListOfBirdiesaysscore": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RealtimeListenResponse"
          },
          {
            "type": "object",
            "description": "Returned by subscribe endpoints with initial data. The client receives the current state\nimmediately and knows which SSE group to listen on for subsequent updates.",
            "additionalProperties": false,
            "properties": {
              "data": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/Birdiesaysscore"
                }
              }
            }
          }
        ]
      },
      "Birdiesaysscore": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "playerName": {
            "type": "string"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      }
    }
  }
}